Moved detectOS() to src/utils/common and adjusted imports
